Use int instead of uint8_t for CoreField::heights_.
CoreField becomes a bit larger, however, this is faster basically.

+++ src/core/core_field.h
@@ -162,7 +162,7 @@ protected:
     template<typename Tracker>
     int dropAfterVanish(int minHeights[], Tracker*);
 
-    uint8_t heights_[MAP_WIDTH];
+    int heights_[MAP_WIDTH];
 };
 
 #endif
